requirements.txt
antlr4-python3-runtime==4.9.3
Bottleneck==1.3.7
cffi==1.16.0
clarabel==0.9.0
ecos==2.0.13
filelock==3.13.1
fsspec==2024.3.1
geotorch==0.3.0
gmpy2==2.1.2
hydra-core==1.3.2
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.4
omegaconf==2.3.0
osqp==0.6.7
packaging==24.0
pandas==2.2.1
pip==24.0
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post2
rpy2==3.5.16
scikit-learn==1.5.0
scipy==1.13.1
scs==3.2.4.post2
setuptools==69.5.1
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.2.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
wheel==0.43.0
